Added the ipython/xkcd-font deploy key, and left a hook for other owners to exist in the same repo. (#22)

+++ .github/commit-and-push-from-travis.sh
@@ -10,10 +10,18 @@ export GIT_AUTHOR_EMAIL=$(git --no-pager show -s --format='%ae' HEAD)
 git status
 
 if [ "${TRAVIS_EVENT_TYPE}" == "push" ] ; then
-    pip install doctr --quiet
-    doctr deploy --built-docs . \
-        --deploy-branch-name ${TRAVIS_BRANCH} . \
-        --key-path .github/deploy_key.enc \
-        --no-require-master \
-        --no-sync
+    # Split the organisation out of the slug. See https://stackoverflow.com/a/5257398/741316 for description.
+    ORG=(${TRAVIS_REPO_SLUG//\// })
+    DEPLOY_KEY=.github/deploy_key.${ORG}.enc
+
+    if [ ! -f ${DEPLOY_KEY} ]; then
+        echo "Not syncing, the deploy key ${DEPLOY_KEY} does not exist"
+    else
+        pip install doctr --quiet
+        doctr deploy --built-docs . \
+            --deploy-branch-name ${TRAVIS_BRANCH} . \
+            --key-path ${DEPLOY_KEY} \
+            --no-require-master \
+            --no-sync
+    fi
 fi
